+++ modules/libcom/src/osi/os/cygwin32/osdNetIntf.c
@@ -1,0 +1,1 @@
+#include "../osdNetIfAddrs.c"
