+++ modules/libcom/src/osi/os/iOS/osdNetIntf.c
@@ -1,0 +1,1 @@
+#include "../osdNetIfAddrs.c"
